task 101 is done

+++ 0x04-more_functions_nested_loops/101-print_number.c
@@ -5,7 +5,7 @@
  * @n: will be printed only using putchar
  */
 
-void print_number(long int n)
+void print_number(int n)
 {
 	int tmp;
